Post, Get, Delete added correctly
This commit adds the ability to start any docker image, get a list of running dockers and delete a running container based on it's name

In the future the return types will have to be converted to JSON to increase the ease of use in the frontend.

In case that a container could not be started, it is falsely reported as successful.
This will also have to be fixed later.